[MOV] point_of_sale, product, website_sale: migrate base unit model
We move up the base unit model from `website_sale` to `product` proper
so that other modules like `point_of_sale` can use it without having
`website_sale` installed.

Task ID: 4277792
