+++ addons/product/__manifest__.py
@@ -38,6 +38,7 @@ Print product labels with barcode.
         'views/res_config_settings_views.xml',
         'views/product_attribute_views.xml',
         'views/product_attribute_value_views.xml',
+        'views/product_base_unit_views.xml',
         'views/product_category_views.xml',
         'views/product_combo_views.xml',
         'views/product_document_views.xml',
