+++ addons/product/models/product_base_unit.py
@@ -1,0 +1,13 @@
+from odoo import fields, models
+
+
+class ProductBaseUnit(models.Model):
+    _name = 'product.base.unit'
+    _description = "Unit of Measure for price per unit on products."
+    _order = 'name'
+
+    name = fields.Char(
+        help="Define a custom unit to display in the price per unit of measure field.",
+        required=True,
+        translate=True,
+    )
