+++ addons/product/models/product_product.py
@@ -140,6 +140,27 @@ class ProductProduct(models.Model):
         store=False,
     )
 
+    base_unit_count = fields.Float(
+        string="Base Unit Count",
+        help="Display base unit price. Set to 0 to hide it for this product.",
+        required=True,
+        default=1,
+    )
+    base_unit_id = fields.Many2one(
+        string="Custom Unit of Measure",
+        help="Define a custom unit to display in the price per unit of measure field.",
+        comodel_name='product.base.unit',
+    )
+    base_unit_price = fields.Monetary(
+        string="Price Per Unit",
+        compute='_compute_base_unit_price',
+    )
+    base_unit_name = fields.Char(
+        help="Displays the custom unit for the products if defined or the selected unit of measure"
+            " otherwise.",
+        compute='_compute_base_unit_name',
+    )
+
     @api.depends('image_variant_1920', 'image_variant_1024')
     def _compute_can_image_variant_1024_be_zoomed(self):
         for record in self:
@@ -487,6 +508,31 @@ class ProductProduct(models.Model):
             product.import_attribute_values = ','.join(sorted(
                 f'{ptav.attribute_line_id.attribute_id.name}:{ptav.product_attribute_value_id.name}'
                 for ptav in product.product_template_attribute_value_ids
+            ))
+
+    def _get_base_unit_price(self, price):
+        self.ensure_one()
+        return self.base_unit_count and price / self.base_unit_count
+
+    @api.depends('lst_price', 'base_unit_count')
+    def _compute_base_unit_price(self):
+        for product in self:
+            if not product.id:
+                product.base_unit_price = 0
+            else:
+                product.base_unit_price = product._get_base_unit_price(product.lst_price)
+
+    @api.depends('uom_name', 'base_unit_id')
+    def _compute_base_unit_name(self):
+        for product in self:
+            product.base_unit_name = product.base_unit_id.name or product.uom_name
+
+    @api.constrains('base_unit_count')
+    def _check_base_unit_count(self):
+        if any(product.base_unit_count < 0 for product in self):
+            raise ValidationError(_(
+                "The value of Base Unit Count must be greater than 0."
+                " Use 0 to hide the price per unit on this product."
             ))
 
     def _load_records_write(self, values):
