+++ addons/product/security/product_security.xml
@@ -30,6 +30,10 @@
 
 <data noupdate="1">
 
+    <record id="group_show_uom_price" model="res.groups">
+        <field name="name">UOM price display</field>
+    </record>
+
     <record id="product_comp_rule" model="ir.rule">
         <field name="name" >Product multi-company</field>
         <field name="model_id" ref="model_product_template"/>
