+++ addons/product/views/product_base_unit_views.xml
@@ -3,7 +3,7 @@
 
     <record id="base_unit_action" model="ir.actions.act_window">
         <field name="name">Base Units</field>
-        <field name="res_model">website.base.unit</field>
+        <field name="res_model">product.base.unit</field>
         <field name="view_mode">list,form</field>
     </record>
