+++ addons/website_sale/models/product_template.py
@@ -132,31 +132,6 @@ class ProductTemplate(models.Model):
         copy=True,
     )
 
-    base_unit_count = fields.Float(
-        string="Base Unit Count",
-        help="Display base unit price on your eCommerce pages. Set to 0 to hide it for this"
-        " product.",
-        compute="_compute_base_unit_count",
-        inverse="_set_base_unit_count",
-        store=True,
-        required=True,
-        default=0,
-    )
-    base_unit_id = fields.Many2one(
-        string="Custom Unit of Measure",
-        help="Define a custom unit to display in the price per unit of measure field.",
-        comodel_name="website.base.unit",
-        compute="_compute_base_unit_id",
-        inverse="_set_base_unit_id",
-        store=True,
-    )
-    base_unit_price = fields.Monetary(string="Price Per Unit", compute="_compute_base_unit_price")
-    base_unit_name = fields.Char(
-        compute="_compute_base_unit_name",
-        help="Displays the custom unit for the products if defined or the selected unit of measure"
-        " otherwise.",
-    )
-
     compare_list_price = fields.Monetary(
         string="Compare to Price",
         help="Add a strikethrough price to your /shop and product pages for comparison purposes."
@@ -224,42 +199,6 @@ class ProductTemplate(models.Model):
         """Set `publish_date` to the moment of (re-)publishing."""
         self.filtered("is_published").publish_date = fields.Datetime.now()
 
-    @api.depends("product_variant_ids", "product_variant_ids.base_unit_count")
-    def _compute_base_unit_count(self):
-        self.base_unit_count = 0
-        for template in self.filtered(lambda template: len(template.product_variant_ids) == 1):
-            template.base_unit_count = template.product_variant_ids.base_unit_count
-
-    def _set_base_unit_count(self):
-        for template in self:
-            if len(template.product_variant_ids) == 1:
-                template.product_variant_ids.base_unit_count = template.base_unit_count
-
-    @api.depends("product_variant_ids", "product_variant_ids.base_unit_count")
-    def _compute_base_unit_id(self):
-        self.base_unit_id = self.env["website.base.unit"]
-        for template in self.filtered(lambda template: len(template.product_variant_ids) == 1):
-            template.base_unit_id = template.product_variant_ids.base_unit_id
-
-    def _set_base_unit_id(self):
-        for template in self:
-            if len(template.product_variant_ids) == 1:
-                template.product_variant_ids.base_unit_id = template.base_unit_id
-
-    def _get_base_unit_price(self, price):
-        self.ensure_one()
-        return self.base_unit_count and price / self.base_unit_count
-
-    @api.depends("list_price", "base_unit_count")
-    def _compute_base_unit_price(self):
-        for template in self:
-            template.base_unit_price = template._get_base_unit_price(template.list_price)
-
-    @api.depends("uom_name", "base_unit_id.name")
-    def _compute_base_unit_name(self):
-        for template in self:
-            template.base_unit_name = template.base_unit_id.name or template.uom_name
-
     def _compute_website_url(self):
         super()._compute_website_url()
         for product in self:
@@ -290,11 +229,6 @@ class ProductTemplate(models.Model):
         return super().write(vals)
 
     # === BUSINESS METHODS ===#
-
-    def _prepare_variant_values(self, combination):
-        variant_dict = super()._prepare_variant_values(combination)
-        variant_dict["base_unit_count"] = self.base_unit_count
-        return variant_dict
 
     def _get_website_accessory_product(self):
         domain = Domain(self.env["website"].sale_product_domain())
@@ -688,7 +622,7 @@ class ProductTemplate(models.Model):
             "taxes": taxes,  # taxes after fpos mapping
         })
 
-        if self.env["res.groups"]._is_feature_enabled("website_sale.group_show_uom_price"):
+        if self.env["res.groups"]._is_feature_enabled("product.group_show_uom_price"):
             price_per_product_uom = uom._compute_price(
                 price=combination_info["price"], to_unit=self.uom_id
             )
