+++ addons/website_sale/security/ir.model.access.csv
@@ -59,11 +59,11 @@ access_ecom_extra_fields_public_public,access_ecom_extra_field public,model_webs
 access_ecom_extra_fields_public_portal,access_ecom_extra_field public,model_website_sale_extra_field,base.group_portal,1,0,0,0
 access_ecom_extra_fields_public_employee,access_ecom_extra_field public,model_website_sale_extra_field,base.group_user,1,0,0,0
 access_ecom_extra_fields_restricted_editor,access_ecom_extra_field restricted_editor,model_website_sale_extra_field,website.group_website_restricted_editor,1,1,1,1
-access_website_base_unit_public_public,website.base.unit public,model_website_base_unit,base.group_public,1,0,0,0
-access_website_base_unit_public_portal,website.base.unit public,model_website_base_unit,base.group_portal,1,0,0,0
-access_website_base_unit_public_employee,website.base.unit public,model_website_base_unit,base.group_user,1,0,0,0
-access_website_base_unit_sale_manager,website.base.unit sale manager,model_website_base_unit,sales_team.group_sale_manager,1,1,1,1
 access_website_checkout_step_website_designer,access_website_checkout_step website_designer,model_website_checkout_step,website.group_website_designer,1,1,1,1
+
+access_product_base_unit_public_public,product.base.unit public public,product.model_product_base_unit,base.group_public,1,0,0,0
+access_product_base_unit_public_portal,product.base.unit public portal,product.model_product_base_unit,base.group_portal,1,0,0,0
+access_product_base_unit_sale_manager,product.base.unit sale manager,product.model_product_base_unit,sales_team.group_sale_manager,1,1,1,1
 
 access_product_feed_base_group_system,access_product_feed base_group_system,model_product_feed,base.group_system,1,1,1,1
 access_product_feed_website_group_website_designer,access_product_feed website_group_website_designer,model_product_feed,website.group_website_designer,1,1,1,1
