+++ addons/website_sale/security/res_groups.xml
@@ -2,10 +2,6 @@
 <odoo noupdate="1">
 
     <!-- Feature groups -->
-    <record id="group_show_uom_price" model="res.groups">
-        <field name="name">UOM Price Display for eCommerce</field>
-    </record>
-
     <record id="group_product_price_comparison" model="res.groups">
         <field name="name">Comparison Price</field>
     </record>
