+++ addons/website_sale/tests/test_google_merchant_center.py
@@ -302,7 +302,7 @@ class TestWebsiteSaleGMC(WebsiteSaleGMCCommon, HttpCase):
     def _setup_6l_water_pack(self):
         self.env.user.group_ids |= self.env.ref("uom.group_uom")
         uom_litre = self.env.ref("uom.product_uom_pack_6")
-        base_unit_litre = self.env["website.base.unit"].create({"name": "L"})
+        base_unit_litre = self.env["product.base.unit"].create({"name": "L"})
         six_pack = self.env["product.product"].create([
             {
                 "name": "Water Pack 6L",
@@ -322,7 +322,7 @@ class TestWebsiteSaleGMC(WebsiteSaleGMCCommon, HttpCase):
         self.assertNotIn("unit_pricing_measure", self.items[six_pack])
 
         # enable "Product Reference Price" setting
-        self.env.user.group_ids |= self.env.ref("website_sale.group_show_uom_price")
+        self.env.user.group_ids |= self.env.ref("product.group_show_uom_price")
         self.update_items()
 
         self.assertEqual("6.0l", self.items[six_pack]["unit_pricing_measure"], "$12 / 6l")
