+++ addons/website_sale/views/res_config_settings_views.xml
@@ -110,7 +110,7 @@
                 <setting
                     string="Product Reference Price"
                     id="ecom_uom_price_option_setting"
-                    help="Add a reference price per UoM on products (i.e $/kg), in addition to the
+                    help="Add a reference price per UoM on products (e.g $/kg), in addition to the
                           sale price"
                     info="This setting is common to all websites."
                 >
